# Conflicts:
#	src/engine/sparqlExpressions/NaryExpression.h
#	test/SparqlExpressionTest.cpp

+++ src/engine/GroupBy.h
@@ -66,7 +66,7 @@ class GroupBy : public Operation {
   virtual float getMultiplicity(size_t col) override;
 
  private:
-  size_t getSizeEstimateBeforeLimit() override;
+  uint64_t getSizeEstimateBeforeLimit() override;
 
  public:
   virtual size_t getCostEstimate() override;
